backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Supabase ───────────────────────────────────────────────────────────────────
supabase

# ── Auth / JWT ─────────────────────────────────────────────────────────────────
python-jose[cryptography]

# ── Environment ────────────────────────────────────────────────────────────────
python-dotenv

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic[email]

# ── HTTP client (used internally by supabase-py) ──────────────────────────────
httpx